Move condition type into its own module
