+++ snuba/datasets/cdc/groupedmessage.py
@@ -7,7 +7,7 @@ from snuba.datasets.dataset_schemas import DatasetSchemas
 from snuba.datasets.cdc.groupedmessage_processor import GroupedMessageProcessor, GroupedMessageRow
 from snuba.datasets.schemas.tables import ReplacingMergeTreeSchema
 from snuba.datasets.table_storage import TableWriter, KafkaStreamLoader
-from snuba.query.query import Condition
+from snuba.query.types import Condition
 from snuba.snapshots.loaders.single_table import SingleTableBulkLoader
 from snuba.util import qualified_column
