+++ snuba/query/types.py
@@ -1,0 +1,6 @@
+from typing import Any, Sequence, Tuple, Union
+
+Condition = Union[
+    Tuple[Any, Any, Any],
+    Sequence[Any],
+]
